[feat] add mapStruct library [feat] product controller & service done

+++ src/main/java/com/haiss/shoppingcart/domain/DTO/PaginationResponse.java
@@ -8,7 +8,7 @@ public class PaginationResponse<T> {
     private List<T> content;
     private int PageNo ;
     private int PageSize ;
-    private int totalElments ;
+    private long totalElments ;
     private int totalPages ;
     private boolean isLast ;
